+++ packages/dsw-data-seeder/CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- New `tablePrefix` option for entries in `extraDatabases` (default `w_`)
+
 
 ## [4.34.0]